qcacld-3.0: Disable EMLSR cap for non-EMLSR cert cases
As part of aux support, the EMLSR capability is set by
default in STA. But, for non-EMLSR WFA certification cases,
this capability needs to be reset. Thus, add logic to
set the EMLSR vendor command flag to true in case of aux
EMLSR support. This solution works for both HMT and GNG
STA for both EMLSR and non-EMLSR certification.

CRs-Fixed: 3833807
Change-Id: I3cbab2d87709793f2a94202efe4f4bb0c7ec441b