qcacld-3.0: Correct the VDEV ID mapping
Currently, driver uses VDEV ID and session ID interchangeably in
the code as in general, VDEV ID and Session are equal. But for some
cases, when VDEV ID is different from the session ID, it will lead
to wrong functionality.
For example, in this case, where VDEV ID is 2 and Session ID is 1
and lim_abort_change_change() maps session ID instead of the
VDEV ID. It leads to assert condition in csr_roam_call_callback()
API. The API tries to check for valid session for given VDEV ID but
as it have session ID instead of VDEV ID, it causes failure in
this condition and hence, it leads to assertion in the code.

So, to fix this, map correct VDEV ID instead of session ID.

Change-Id: Ib66cb5df6ca1f7e03d2824116d579227e8ba04b6
CRs-Fixed: 3802069