qcacld-3.0: Consider the channel width in OMN IE during assoc
If OMN IE is present in the assoc response, the host driver
doesn't downgrade the channel width, leading to unnecessary
retries on higher bandwidth.

To fix this, always consider the OMN IE during the VHT channel
width computation. Restrict the VHT channel width based on the
OMN IE channel width.
If the OMN IE is present in the assoc response, then set the
channel width and phymode in peer assoc command based on it.

Change-Id: Ib7d3ebba73eb645425f4da50a6602192d927a9d5
CRs-Fixed: 3741156