qcacld-3.0: Pass correct mac address to enable/disable powersave
Currently in the function __wlan_hdd_cfg80211_set_power_mgmt(),
if the adapter is a ML adapter, driver passes link mac address
to the API which enables/disables powersave or adapter mac address
in case of legacy addapter. The issue with this approach is, if
the adapter is a ML adapter and the connection is a non-ML
legacy connection, driver passes wrong mac address leading to a
failure in sending start/stop dhcp indication to FW.

To resolve this issue, pass correct mac to the API by checking
the type of connection and not by adapter type.

Change-Id: I03e681fecb162a80cdd0febd824fca8fa1d5c3fd
CRs-Fixed: 3833102