qcacld-3.0: Cache link specific reassoc rsp in roam join_rsp
GETBSSINFO queries for the link info on which the disconnect
is processed. When disconnect is issued on non-assoc link
post roaming, it returns invalid stats as it fetches the
stats from assoc link assoc response, where some IEs could
be different between assoc link(e.g 5 GHz link) and partner
link(2 GHz link). In current issue, bandwidth is reported
incorrect.
Currently, link specific reassoc response is generated
and cached in pe_session. This is used in lim but it's
not propagated to connection mgr and other upper layers
as there is no consumer so far. Link specific reassoc
response is needed in connection mgr as well to serve
stats requests(e.g. cached reassoc rsp is used for
QCA_NL80211_VENDOR_SUBCMD_GET_STATION).
Generate and cache the link specific reassoc response when
ROAM_SYNC is received inorder to serve this.

Change-Id: I0a56fc6648b28d447dd6bfd60f3642940abce833
CRs-Fixed: 3831067