qcacld-3.0: Fix address translation check for roaming
During address translation of frames from MLD -> Link address
or vice versa, then the CM vdev state is checked if roaming
in progress. But the state transition will happen only in
non-wow mode where firmware indicates Roam start to host.
But in wow mode the ROAM_START indication won't be sent by the
firmware and directly roam preauth start event will come.

Fix the check to detect if we are in roaming by checking if
we are already in connected state

CRs-Fixed: 3789966
Change-Id: Ifa1151547d001be21c329068804562349c41bae7