qcacld-3.0: Restrict CSA channel width on DOT11 mode
Currently, STA do CSA on the given channel width, provided by the
AP. But sometimes in the fault conditions, AP provide channel width
which is not supported by STA.
For example, in this case, STA is connected with AP in 11AX mode.
During CSA, AP provides the 320 MHz channel width. But STA supports
maximum 160 MHz in 11AX mode and thus, it leads to disconnection.

So, to fix this, add check for dot11mode and restrict the channel
width to the maximum allowed channel width for the corresponding
dot11mode.

Change-Id: I29593db740afe6b23ea334bb499c46377868cd9b
CRs-Fixed: 3848635