qcacld-3.0: Disable 2 link force active command for Single MAC solution
For the single mac solution, the Force link active command
is allowed to process when performed on more than one link.

Add check during processing of force link active command
if the command instructs to bring a single link to active
state or two link to active state when neither of the two link
is connected to 2.4 GHz band.

Change-Id: I4f867961746841bdba82a4b57c9a921a4b03911e
CRs-Fixed: 3851175