qcacld-3.0: Add SAP movement in case of STA disconnect
Move SAP to NAN social channel on STA disconnect.
If STA and SAP are doing SCC on a NAN non-social channel while NAN is on
a social channel, SAP will do MCC with NAN after STA disconnect.

This change is to mitigate this MCC issue, move SAP to
NAN channel on STA disconnect.

Change-Id: Iabd4244909c9fd9d19a4e98c2dc20c841d970c29
CRs-Fixed: 3852840