qcacld-3.0: NULL check for STC handle when handling flow stats policy
Flow stats policy can be configured even if STC is not enabled.
Currently there is no NULL check in the flow stats policy
handler, which leads to invalid memory access when STC is disabled
and STC handle is NULL.

To fix this issue, check the validity of STC handle before
attempting to configure the flow stats policy.

Change-Id: I133d3d38aada8c23a42f0ddf7b2d349cc03bdbea
CRs-Fixed: 3853663