qcacld-3.0: Handle BS state machine events correctly
Currently host driver does not handle BS state machine events
correctly which may lead to bearer switch request reject from
host driver.
In current issue, host driver receives connect request it sends
BS to BLE request to user space and increments the ref count.
On connect complete, it decrements the ref count and makes it 0
and starts a timer of 10 seconds to switch the bearer back to
WLAN, now before this timer expires, user space requests for BS
to wlan and host driver accepts this request and user space
contiunes the BS to wlan. If the host driver gets disconnect
and connect request again within this 10 seconds, it again
increments the ref count and cancels the timer and wants to
stay in non-wlan bearer. Since there is already BS to wlan is
in progress, once this switch gets complete, host driver gets
an indication of BS to wlan completion and it moves its state
to WLAN bearer, which is not correct as it wants to be in non-wlan
state. Now on connect complete, host driver tries to switch the
bearer back to WLAN and tries to decrement the counter. Since the
state of the BS state machine is already WLAN, it does not
proceed with the bearer switch and also does not decrement the
reference count. This reference count remains forever and never
allows the bearer move back to wlan.
To address above issue, add below changes in the BS state machine:
1. Reject the user space BS request to switch to wlan if the timer
state is running.
2. Reduce the timer value from 10 seconds to 5 seconds.
3. If ref count is non-zero, do not move BS state machine to
wlan.

Change-Id: I6076f67d51726e041c189d7172b661445c9cf730
CRs-Fixed: 3842893