qcacld-3.0: Add TWT responder support for LL SAP
Currently, there is "twt_responder" INI to enable and disable the
TWT responder.
Driver intersects the TWT responder capability of the target with
this INI and configuration parameter from the userspace, and decide
whether TWT responder should be enable or disable. But some targets
requires TWT responder enable for LL SAP only.

So, to support this, introduces new INI "twt_responder_type"
which determines TWT responder requires for SAP or LL_SAP or both.

Change-Id: I321344f85929fa1dc9de9c282787f47612c3dcc0
CRs-Fixed: 3839562