qcacld-3.0: Add action OUI to disable Aux listen
AUX listen is prone to ITO issues. Thus. add a new action
OUI to ensure AUX listen can be disabled if we detect some
ITO issue with certain APs. The cfg item will be present in
the host and sent to the FW to disable Aux listen.
Currently, there are no known APs to add in the OUI string,
but as and when they are identified, their OUIs should be
added in the defined INI.

CRs-Fixed: 3873157
Change-Id: I5cbdfbb7f47f3c181d606083dacec94bb0135b9e