qcacld-3.0: Update opMode member type as enum phy_ch_width
The opMode of tUpdateVHTOpMode is actually passed as enum phy_ch_width
in recent change, but it directly pass to target as
wmi_host_channel_width, it will not same value if BW is 320 MHz.

It is prefer to use enum phy_ch_width as type of opMode, and change
legacy opmode change code to make it compatible with it.
Convert the channel width to wmi_host_channel_width when pass to
target.

Change-Id: Ib269a6e6de161d2eca301b6d54d3472d2c5ab3b7
CRs-Fixed: 3855090