qcacld-3.0: Remove incorrect feature flag for tsf timeout
Whenever host driver wants to do CSA for LL_LT_SAP,
it requests TSF values from FW and starts a timer for the response.
If the response comes or timeout happens, host driver continues with
the vdev restart sequence.
Currently host driver handles this timer handler in
WLAN_FEATUTE_LL_LT_SAP_CSA feature flag. This feature flag is not
defined in host driver and because of this, host does not handle
timeout scenario which leads to CSA failure on LL_LT_SAP with
CSA stuck for vdev start.

To address this issue, remove WLAN_FEATUTE_LL_LT_SAP_CSA feature
flag so that timer handler is processed successfully.

Change-Id: I7e707c0795a79475f95cbcd96fa86c47c5c34795
CRs-Fixed: 3868133