qcacld-3.0: Fix OOB when parse beacon during roaming
When parsing beacon frames during roaming, there is a discrepancy between
the frame length (frame len) and the frame pointer (frame ptr).
Specifically, the frame ptr excludes the fixed parameters part, while the
frame len does not. As a result, the frame len ends up being larger than
the actual Information Element (IE) length, leading to an out-of-bounds
(OOB) situation.

To address this issue, we need to correct the frame len based on the frame
ptr of the beacon.

Change-Id: I90a162ef0d457e0b5038b6522ebafe08768a82aa
CRs-Fixed: 3870592