qcacld-3.0: Don't update max_tx_power as 0 to FW
Currently, if power_constraint_abs is 0 and tpe_ie is
not advertsied by AP or skip_tpe_consideration is false
then diff of regulatory power and local_constraint is
updated to FW as Tx power which may cause issue in case both
powers are same.

Fix is set regulatory tx power in case regulatory tx power
and local constarint is same.

Change-Id: If897684d33f163ae7c879122f1c7be90a58412cd
CRs-Fixed: 3726862