qcacld-3.0: Block NAN + SBS ML-STA concurrency
Currently, NAN discovery is not supported in SBS hw_mode and
forces the hw_mode to be DBS. Thus MCC is not allowed between
2 GHz NAN discovery + 5 GHz link.
When ML-STA is on 5L+5H and NAN discovery is enabled on 2 GHz,
both STA links would be moved to 5 GHz MAC. But MCC between
ML-links is not allowed and host driver needs FORCE_INACTIVE
one link in such cases.
Send MLO_LINK_FORCE_MODE_INACTIVE_NUM with num_of_link=1 to avoid
this.
Add the below supported changes to address this,
1. Handle NAN_DISC also as legacy connection as part of
   nan_pre_enable and consider both the discovery channels(ch-6
   and ch-149).
2. Currently, get_disallow_mlo_mode_table() fetches the NAN state
   from policy_mgr connection table. But connection table doesn't
   contain any entry while NAN is coming up. Entry gets added
   after NAN is UP. So, refine it to check discovery state instead
   of policy_mgr entry.
3. Consider NAN as concurrency for disallowed mode bitmap
   preparation by adding ML_EMLSR_DISALLOW_BY_CONCURENCY to
   add MLMR disallow mode also in unsupported concurrency cases.

Change-Id: I4ddf98d7370bfe7dbe389d70a0d31cadbb188d5f
CRs-Fixed: 3859356