qcacld-3.0: Check high_tput_vote to allow/prevent suspend
Currently, there is a race condition in STA+SAP concurrancy when
if both are getting disconnected simultaneously, vote in
rtpm_tput_policy gets decremented twice as curr_state member
is not updated in other context which is resulting in assert.
To fix this, set/clear high_tput_vote in atomic context so when
multiple contexts try to simultaneously change vote, they get
serialized.

CRs-Fixed: 3875117
Change-Id: Ide0d0fa83f184520feae0146c06096a0fabc6794