qcacld-3.0: Do not take wakelock for LL_LT_SAP client connected case
Currently host driver takes wakelock to prevent system going into
runtime_pm if any client is connected to the SAP. Because of this,
device does not enter into runtime_pm. For LL_LT_SAP, it is high
possibility that clients are connected and in idle state. This leads
to more power consumption.

To fix this issue, do not prevent runtime_pm for LL_LT_SAP even
if any client is connected.

Change-Id: I01870c3928afe69a3ebfdb710d362e1b7deecf93
CRs-Fixed: 3859006