qcacld-3.0: Allow SAP on 2.4 GHz in LL_LT_SAP + non-sap concurrency
Currently host driver does not add 2.4 GHz channels in PCL for SAP
if LL_LT_SAP + non-sap concurrency is present in SBS mode. This leads
to SAP failure issue where if non-sap (for example STA) interface is
present on indoor channel and SAP is not allowed to come up on indoor
channels. In this case, as SAP can not cme up in SCC on indoor channel,
SAP turn on will fail.

To fix this issue, add 2.4 GHz channels as well in SAP PCL so that it
will allow SAP on 2.4 GHz channels. FW will change the HW mode to
DBS when SAP comes up on 2.4 GHz

Change-Id: I75ad683e195f36e67e70bca8330879652c02bb2c
CRs-Fixed: 3870039