qcacld-3.0: Don't avoid rejected link as partner link
As per current implementation If AP rejects association
with code "unable to handle new sta", any candidate for
which the rejected link is partner link will not be
considered for association and connection will happen
in SLO instead of 2 link.

This fix is to not to avoid link as partner link and
connect with 2 link by changing assoc link.

Change-Id: I162aaca09a4a0486aab7807cd4bfa48bc4ef4e6e
CRs-Fixed: 3880009