qcacld-3.0: Re-order cached BS requests
Currently host driver does not re-order cached BS requests which
may lead to issues where response or timeout of one request
can be considered for the other request.
For example BS request for non-wlan is cached at index 0, while
this request is under process, request to switch the bearer to
wlan is received and cached at index at 1. once non-wlan request
is processed index 0 will be freed. Now again if BS request to
non-wlan is received, it will be cached at index 0. BS request
of wlan is already in process which is cached at index 1 and if
this request gets time out then in the timeout handler, host
driver gets the first valid request from the cached requests
and processes the timeout for that request, currently timeout
happens for wlan BS request and host driver treats this timeout
for non-wlan BS request as this is the first valid cached request.
Because of this BS state machine is stuck and not able to
proceed further.

To address above issue, reorder the cached BS requests in such a
way that there is no empty place in the cached array before any
cached request.

Change-Id: I64e18f0919e5d260e2f38cfada2750a8675306d4
CRs-Fixed: 3878784