qcacld-3.0: Update 24 GHz cb correctly
Currently host is not validating that 2 GHz support
is forced to 20 MHz or not while populating he caps
for per sta profile. If partner is 2 GHz and 2 GHz support
is forced to 20 MHz then chan width 0 is not reset which
means that it can support 20 & 40 MHz both.
On reset 24 GHz bonding mode is set to 1 without checking
the value set by ini.

Fix is to check whether force_20mhz_in_24ghz is enabled
or not before updating 2 GHz cb and to reset 24 GHz bonding
mode to ini value gChannelBondingMode24GHz instead of 1.

Change-Id: Ia21b4b85bf6bf8c006ebb3e449c7c2f18664d8d0
CRs-Fixed: 3863351