qcacld-3.0: Switch bearer to non-wlan on LL_LT_SAP stop
Currently host driver does not switch the bearer to non-wlan
when stop_ap is received for ll_lt_sap. Since LL_LTSAP is going
down, it will lead to disconnection with the clients and there
could be gitches observed in the streaming.
Add changes to switch the bearer to non-wlan on stop_ap so that
driver gives an early indication to user space to switch the
bearer to minimise the glitches in the streaming.

Change-Id: I237aa0c2ed7e9e539bb1ee548d35458fd44d50a4
CRs-Fixed: 3890732