qcacld-3.0: Update partner link channel for MBSSID case
In the MBSSID case, the `lim_update_mlo_mgr_ap_link_info_mbssid_connect`
API updates the channel for association and partner links. However, the
channel parameters, such as phymode, are not updated in the MLO link
info, causing the firmware to receive an invalid phymode.

This fix calls the `lim_update_mlo_mgr_info` API to update the link
channel parameter info in the MLO manager data structure.

Change-Id: I71aecd04f6aed6481693e291aab11175b7f42e84
CRs-Fixed: 3881093