qcacld-3.0: Update check to terminate ACS early in 2 GHz
In case of 2 GHz early termination, only non-overlapping channel
should be considered for early termination. Currently the check
for non-overlapping is not there resulting in selection of
overlapping channel for early termination in case no beacon/probe
response is received on that channel. Adding the check for the same.

Change-Id: I5e8f2d24d694b88f76b4f1da4c1f864f623e9165
CRs-Fixed: 3895340