qcacld-3.0: Do not disable roaming on STA1 upon STA2 connect
Currently in STA+STA case, roaming is disabled on first STA interface when connect request is received on second STA interface but this results in ROAM_ABORT if roaming has already started on STA1, which may lead to HO_FAIL.
Currently this is skipped if ROAM_SYNC is in progress on STA1 but ROAM_START case is not considered for this. Though host hasn't got ROAM_SYNC yet, firmware might be in the middle of roaming and it's better to avoid causing HO_FAIL.

Fix is to avoid disabling roaming if ROAM START is received on first STA interface when connect request is received on other STA interface.

CRs-Fixed: 3897546
Change-Id: I4950191f7d94e9bb37541ac0d183ac638cd9ffea