qcacld-3.0: Fix ACS logic to select 320 MHz SAP channel
1) In the 320 MHz sap channel sorting API, the AP's best
power type has to be fed to reg set channel params API and
not the hardcoded macro.
2) There are multiple overlapping bonded pairs for 320 MHz,
and currently if 1st channel weight is computed, rest of 15
channels are all assigned max weight and marked as
calculated. So, ACS will not consider next frequencies.
For e.g., 6135 is part of both 5955-6255 as well as
6115-6415 bonded pairs but due to current logic, 6135 is not
considered for 2nd bonded pair, so are the subsequent ones.

Since 6100 is vlp cutoff frequency for US, 6135 or higher
frequency must be preferred over 5975, hence the weight
calculation done flag computation is updated. This helps to
correctly choose SAP channels for VLP mode.

CRs-Fixed: 3892569
Change-Id: Icc4e7d30cb478d48cabe7bbb64b1cfa50d0574ad