qcacld-3.0: check standby to active link transition before STA+STA check
Single phy solution does not broadcast multiple STA support. As a
result in api policy_mgr_allow_sta_concurrency(), rejection is
observed when link switch occurs in a MLSR connection on a 11BE
single phy solution.

Modify the checks to allow link switch when the current connection
is a MLO connection. If the current connection is a legacy
connection then check for multiple STA support.

Change-Id: Iea660b2488382d22596cea9abfd0b9c600fca7bb
CRs-Fixed: 3900548