qcacld-3.0: Fix inactivity check for classified flow
Currently, for the classified flows, there is an
inactivity check done by checking the last access
timestamp for TX and RX flow for the classified stream.

The current implementation marks a stream as inactive
even if one of the TX or RX flow is inactive. In such
scenario, if one among the TX/RX flow is inactive and
the other flow is active, the classified stream gets
marked as inactive-active consecutively forever.
This will lead to sending of the active traffic map
indication to FW constantly.

To fix the above mentioned issue, mark a classified
stream as inactive only if both its TX and RX flows
are inactive.

Change-Id: I17a3ed6a9944f2db8cfe9d3808fa414c0f43dd86
CRs-Fixed: 3909455