qcacld-3.0: Ensure Wiphy Lock is Held When Calling Kernel Wiphy sync API
The kernel expects the wiphy lock to be held when calling wiphy sync API,
such as regulatory_set_wiphy_regd_sync. Currently, this is not being done,
leading to kernel assertions when updating the wiphy regulatory settings.

To prevent kernel assertions, ensure the wiphy lock is held when the driver
calling kernel wiphy sync APIs.

Change-Id: I9b9ba26b8bda8f1a5cd82c53f6fe6b9b136e8dcd
CRs-Fixed: 3908451