qcacld-3.0: Fix VDEV stop timeout for monitor mode
Host driver waits for vdev stop response even if the
monitor mode interface is just created and not initialized.
This leads to timeout, since driver does not send vdev stop
command to firmware.

To fix this, reset the wait for completion if the vdev moves
to DOWN state directly.

Change-Id: Idf50464ac680e07b54a850e94e9d5c91ac6b6317
CRs-Fixed: 3910287