qcacld-3.0: Free memory in caller to avoid use-after-free
Driver saves the assoc req frame received from each SAP clients
and cleans this memory either if there is error in successfully
adding the client or when client disconnects. This memory is
allocated in lim_process_assoc_req_frame() and call to
lim_proc_assoc_req_frm_cmn() API receives this memory pointer
as function argument. Incase of any error, the callee API clears
the memory allocated as part of error handling and returns error
status, caller is currently not checking the return status and
accesses the same memory pointer which will result in use-after-free.

Let the API where the memory gets allocated take care of cleanup so
that the cleanup ensures no invalid memory access and also check
the return status of callee API to branch out to error handling
in the caller.

CRs-Fixed: 3910741
Change-Id: I470463bd39d0d128c0fd7cfc1e22a75a6140145c