qcacld-3.0: Add capability APIs to use sta vdev for p2p device
Add APIs to get firmware capability whether it supports to use
STA vdev for P2P or not.

Define two INIs to use sta vdev for p2p device operations,
1. sta_vdev_for_p2p_device: This INI is used to use sta vdev
   when p2p device interface is brought UP. No vdev gets created
   for P2P-device mode and all p2p device operations are going
   to use STA vdev.

2. sta_vdev_for_p2p_device_upon_vdev_exhaust: This INI is used
   to use the STA vdev for P2P-device operations in vdev
   exhaust cases.
   Currently, host driver creates a vdev(s) for every interface
   UP. When all the vdevs are consumed, it fails to create the
   vdev for new interface that is coming UP. Check if vdev is
   created for P2P-device when such vdev exhaust happens and
   delete that vdev to accommodate vdev for the new interface.

Define an API to set a flag when firmware and host driver
INI param support the feature and STA vdev is available.

Change-Id: I238ccd687c1360709388851b78b9efacc0359a19
CRs-Fixed: 3808087