qcacld-3.0: Fetch Max wmi buffer length to evaluate roam stats event
For the roam stats event, The maximum expected data is ~1500 bytes.
This results in sanity check for the event data when the
event data received exceeds the maximum expected data limit.

Evaluate the roam stats event data with the wmi max message length
instead of an absolute value to prevent the data sanity check failure

Change-Id: Ia5fc747ec6a7e1987eb717cb4f366323a8fa7162
CRs-Fixed: 3916459