qcacld-3.0: Update set_tpc_power to fw post roam sync
Currently, host driver computes tpc power from the
beacon/probe rsp received in roam_sync_ind and updates
it to fw as soon as it computes.
But this sends a WMI command to fw while roam sync
handling is in progress and it causes a delay
(around 5 ms) in handling the ROAM_SYNC_COMPLETE
command by fw.
So, compute the tpc power from beacon/probe rsp as
part of roam sync handling and update the computed
tpc power to fw post ROAM_SYNC_COMPLETE.

CRs-Fixed: 3872867
Change-Id: I359ae8410e5a8307b96b139d416a669e7ba39b50