qcacld-3.0: Cache Roam Policy to avoid setting of duplicate roam policy
Observe multiple instances when same roam policy is being set from
the framework resulting in resetting of RSO state machine in
framework. Due to this behavior multiple instance of
consecutive partial roam scan is observed.

Cache roam policy received from the framework in order to reject
new roam policy received from the framework if the new received
roam policy is same as the roam policy that is in effect. Revert
the cached roam policy with WLAN_ROAMING_NOT_ALLOWED when
restoring the default roam configuration in order to not
interfere while applying the new roam policy that driver may receive
after restoring the roam configuration.

Change-Id: I4ffacce7efc385238ceb4315238a76f1076f3527
CRs-Fixed: 3905124