qcacld-3.0: Reorder VSIEs to end of beacon or probe rsp temp
Driver uses auto parser to pack Vendor IEs in beacon or probe response
template and appends the manually packed IEs at the end of the buffer.
Some commercially available client devices fail to scan 6 GHz channels
when Vendor specific IEs occur in the middle of the IEs and results in
DOS failure.

Repack the beacon and probe response template to reorder Vendor IEs
to the end of the template.

Change-Id: I06531313532d2bc0bf156b1f2ea30dd268a8ac6b
CRs-Fixed: 3902571