qcacld-3.0: Use original p2p0 wdev instead of a copy for p2p rx
Currently, a copy of the p2p0 wdev is made whenever there is an
rx frame and it's used to send the frame to kernel. The copied
wdev might go out of sync from the original wdev when original
wdev entries(e.g. mgmt_registrations) get modified in user
thread context. It may lead to node corruption when one thread
is setting it to NULL and other thread is copying from the
same node.
Use original wdev instead of a copy to avoid out of sync/usage
of deleted entries as the original wdev entries are always
accessed with a lock(e.g. mgmt_registrations_lock).

Also, all P2P actions frames are allowed when P2P-GO is
running currently. Driver is supposed to allow frames to
let new P2P peers to join the group. So, allow only P2P
discovery frames on when ROC is not running to avoid sending
unwanted frames.

Change-Id: I5fecc5e16627be7fe2c14630cb67fc40af27ab9b
CRs-Fixed: 3935355